fix linters bug
